fixed cost in order item, it was not displaying cents

+++ database_seed.sql
@@ -90,7 +90,7 @@ create table order_item
     user_id   int          not null,
     item_id   int          not null,
     item_name varchar(255) not null,
-    cost      float(11, 2) not null,
+    cost      decimal(11,2) not null,
     qty       int          not null,
     primary key (order_id, user_id, item_id)
 );
